refactor: resolve circular deps with sequencer type exports
