+++ packages/sequencer/src/helpers/query/BlockExplorerTransportModule.ts
@@ -1,6 +1,8 @@
-import { BlockModel } from "@proto-kit/api";
+import { Block } from "../../storage/model/Block";
+
+export type ClientBlock = Pick<Block, "hash" | "previousBlockHash" | "height" | "transactions" | "transactionsHash">;
 
 export interface BlockExplorerTransportModule {
   waitTxInclusion(txHash: string): Promise<string>;
-  getBlock(param?: string | number): Promise<BlockModel | undefined>;
+  getBlock(param?: string | number): Promise<ClientBlock | undefined>;
 }
